# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ujraa API Documentation",
        "description": "API documentation for Ujraa platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/jobs/{job}/proposals": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job proposals",
                "operationId": "1ea5a53b1fd4b1db76cdfb7b03e2c6d3",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of proposals for the job"
                    },
                    "403": {
                        "description": "Not authorized to view proposals"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Submit a proposal for a job",
                "operationId": "ad1e863f8b1d41f09574fc621863d030",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "delivery_days",
                                    "cover_letter"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "delivery_days": {
                                        "type": "integer"
                                    },
                                    "cover_letter": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proposal submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "delivery_days": {
                                                    "type": "integer"
                                                },
                                                "cover_letter": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "credits_used": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not enough credits or not authorized"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/proposals/{proposal}/accept": {
            "post": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Accept a proposal",
                "operationId": "fc3e4fa7f6761c262a6a9b6e4f665c68",
                "parameters": [
                    {
                        "name": "proposal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposal accepted successfully"
                    },
                    "403": {
                        "description": "Not authorized to accept this proposal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/proposals/{proposal}/reject": {
            "post": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Reject a proposal",
                "operationId": "fbd6f95be9243c5e8de682ce2455fa8e",
                "parameters": [
                    {
                        "name": "proposal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposal rejected successfully"
                    },
                    "403": {
                        "description": "Not authorized to reject this proposal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}